+++ lib/src/hooks.dart
@@ -1,4 +1,132 @@
-part of 'hook.dart';
+part of 'framework.dart';
+
+/// A [HookWidget] that defer its [HookWidget.build] to a callback
+class HookBuilder extends HookWidget {
+  /// The callback used by [HookBuilder] to create a widget.
+  ///
+  /// If a [Hook] asks for a rebuild, [builder] will be called again.
+  /// [builder] must not return `null`.
+  final Widget Function(BuildContext context) builder;
+
+  /// Creates a widget that delegates its build to a callback.
+  ///
+  /// The [builder] argument must not be null.
+  const HookBuilder({
+    @required this.builder,
+    Key key,
+  })  : assert(builder != null),
+        super(key: key);
+
+  @override
+  Widget build(BuildContext context) => builder(context);
+}
+
+/// A state holder that allows mutations by dispatching actions.
+abstract class Store<State, Action> {
+  /// The current state.
+  ///
+  /// This value may change after a call to [dispatch].
+  State get state;
+
+  /// Dispatches an action.
+  ///
+  /// Actions are dispatched synchronously.
+  /// It is impossible to try to dispatch actions during [HookWidget.build].
+  void dispatch(Action action);
+}
+
+/// Composes an [Action] and a [State] to create a new [State].
+///
+/// [Reducer] must never return `null`, even if [state] or [action] are `null`.
+typedef Reducer<State, Action> = State Function(State state, Action action);
+
+/// An alternative to [useState] for more complex states.
+///
+/// [useReducer] manages an read only state that can be updated
+/// by dispatching actions which are interpreted by a [Reducer].
+///
+/// [reducer] is immediatly called on first build with [initialAction]
+/// and [initialState] as parameter.
+///
+/// It is possible to change the [reducer] by calling [useReducer]
+///  with a new [Reducer].
+///
+/// See also:
+///  * [Reducer]
+///  * [Store]
+Store<State, Action> useReducer<State extends Object, Action>(
+  Reducer<State, Action> reducer, {
+  State initialState,
+  Action initialAction,
+}) {
+  return Hook.use(_ReducerdHook(reducer,
+      initialAction: initialAction, initialState: initialState));
+}
+
+class _ReducerdHook<State, Action> extends Hook<Store<State, Action>> {
+  final Reducer<State, Action> reducer;
+  final State initialState;
+  final Action initialAction;
+
+  const _ReducerdHook(this.reducer, {this.initialState, this.initialAction})
+      : assert(reducer != null);
+
+  @override
+  _ReducerdHookState<State, Action> createState() =>
+      _ReducerdHookState<State, Action>();
+}
+
+class _ReducerdHookState<State, Action>
+    extends HookState<Store<State, Action>, _ReducerdHook<State, Action>>
+    implements Store<State, Action> {
+  @override
+  State state;
+
+  @override
+  void initHook() {
+    super.initHook();
+    state = hook.reducer(hook.initialState, hook.initialAction);
+    assert(state != null);
+  }
+
+  @override
+  void dispatch(Action action) {
+    final res = hook.reducer(state, action);
+    assert(res != null);
+    if (state != res) {
+      setState(() {
+        state = res;
+      });
+    }
+  }
+
+  @override
+  Store<State, Action> build(BuildContext context) {
+    return this;
+  }
+}
+
+/// Create and cache the instance of an object.
+///
+/// [useMemoized] will immediatly call [valueBuilder] on first call and store its result.
+/// Later calls to [useMemoized] will reuse the created instance.
+///
+///  * [keys] can be use to specify a list of objects for [useMemoized] to watch.
+/// So that whenever [Object.operator==] fails on any parameter or if the length of [keys] changes,
+/// [valueBuilder] is called again.
+T useMemoized<T>(T Function() valueBuilder, [List keys = const <dynamic>[]]) {
+  return Hook.use(_MemoizedHook(
+    valueBuilder,
+    keys: keys,
+  ));
+}
+
+/// Obtain the [BuildContext] of the currently builder [HookWidget].
+BuildContext useContext() {
+  assert(HookElement._currentContext != null,
+      '`useContext` can only be called from the build method of HookWidget');
+  return HookElement._currentContext;
+}
 
 class _MemoizedHook<T> extends Hook<T> {
   final T Function() valueBuilder;
@@ -25,6 +153,14 @@ class _MemoizedHookState<T> extends HookState<T, _MemoizedHook<T>> {
   T build(BuildContext context) {
     return value;
   }
+}
+
+/// Watches a value.
+///
+/// Whenever [useValueChanged] is called with a diffent [value], calls [valueChange].
+/// The value returned by [useValueChanged] is the latest returned value of [valueChange] or `null`.
+R useValueChanged<T, R>(T value, R valueChange(T oldValue, R oldResult)) {
+  return Hook.use(_ValueChangedHook(value, valueChange));
 }
 
 class _ValueChangedHook<T, R> extends Hook<R> {
@@ -54,6 +190,21 @@ class _ValueChangedHookState<T, R>
   R build(BuildContext context) {
     return _result;
   }
+}
+
+/// Create  value and subscribes to it.
+///
+/// Whenever [ValueNotifier.value] updates, it will mark the caller [HookWidget]
+/// as needing build.
+/// On first call, inits [ValueNotifier] to [initialData]. [initialData] is ignored
+/// on subsequent calls.
+///
+/// See also:
+///
+///  * [ValueNotifier]
+///  * [useStreamController], an alternative to [ValueNotifier] for state.
+ValueNotifier<T> useState<T>([T initialData]) {
+  return Hook.use(_StateHook(initialData: initialData));
 }
 
 class _StateHook<T> extends Hook<ValueNotifier<T>> {
@@ -90,15 +241,27 @@ class _StateHookState<T> extends HookState<ValueNotifier<T>, _StateHook<T>> {
   }
 }
 
-class _TickerProviderHook extends Hook<TickerProvider> {
-  const _TickerProviderHook([List keys]) : super(keys: keys);
+/// Creates a single usage [TickerProvider].
+///
+/// See also:
+///  * [SingleTickerProviderStateMixin]
+TickerProvider useSingleTickerProvider({List keys}) {
+  return Hook.use(
+    keys != null
+        ? _SingleTickerProviderHook(keys)
+        : const _SingleTickerProviderHook(),
+  );
+}
+
+class _SingleTickerProviderHook extends Hook<TickerProvider> {
+  const _SingleTickerProviderHook([List keys]) : super(keys: keys);
 
   @override
   _TickerProviderHookState createState() => _TickerProviderHookState();
 }
 
 class _TickerProviderHookState
-    extends HookState<TickerProvider, _TickerProviderHook>
+    extends HookState<TickerProvider, _SingleTickerProviderHook>
     implements TickerProvider {
   Ticker _ticker;
 
@@ -135,6 +298,41 @@ class _TickerProviderHookState
     if (_ticker != null) _ticker.muted = !TickerMode.of(context);
     return this;
   }
+}
+
+/// Creates an [AnimationController] automatically disposed.
+///
+/// If no [vsync] is provided, the [TickerProvider] is implicitly obtained using [useSingleTickerProvider].
+/// If a [vsync] is specified, changing the instance of [vsync] will result in a call to [AnimationController.resync].
+/// It is not possible to switch between implicit and explicit [vsync].
+///
+/// Changing the [duration] parameter automatically updates [AnimationController.duration].
+///
+/// [initialValue], [lowerBound], [upperBound] and [debugLabel] are ignored after the first call.
+///
+/// See also:
+///   * [AnimationController]
+///   * [useAnimation]
+AnimationController useAnimationController({
+  Duration duration,
+  String debugLabel,
+  double initialValue = 0,
+  double lowerBound = 0,
+  double upperBound = 1,
+  TickerProvider vsync,
+  AnimationBehavior animationBehavior = AnimationBehavior.normal,
+  List keys,
+}) {
+  return Hook.use(_AnimationControllerHook(
+    duration: duration,
+    debugLabel: debugLabel,
+    initialValue: initialValue,
+    lowerBound: lowerBound,
+    upperBound: upperBound,
+    vsync: vsync,
+    animationBehavior: animationBehavior,
+    keys: keys,
+  ));
 }
 
 class _AnimationControllerHook extends Hook<AnimationController> {
@@ -205,6 +403,36 @@ Switching between controller and uncontrolled vsync is not allowed.
   }
 }
 
+/// Subscribes to a [ValueListenable] and return its value.
+///
+/// See also:
+///   * [ValueListenable]
+///   * [useListenable], [useAnimation], [useStream]
+T useValueListenable<T>(ValueListenable<T> valueListenable) {
+  useListenable(valueListenable);
+  return valueListenable.value;
+}
+
+/// Subscribes to a [Listenable] and mark the widget as needing build
+/// whenever the listener is called.
+///
+/// See also:
+///   * [Listenable]
+///   * [useValueListenable], [useAnimation], [useStream]
+void useListenable(Listenable listenable) {
+  Hook.use(_ListenableHook(listenable));
+}
+
+/// Subscribes to an [Animation] and return its value.
+///
+/// See also:
+///   * [Animation]
+///   * [useValueListenable], [useListenable], [useStream]
+T useAnimation<T>(Animation<T> animation) {
+  useListenable(animation);
+  return animation.value;
+}
+
 class _ListenableHook extends Hook<void> {
   final Listenable listenable;
 
@@ -242,6 +470,15 @@ class _ListenableStateHook extends HookState<void, _ListenableHook> {
     super.dispose();
     hook.listenable.removeListener(_listener);
   }
+}
+
+/// Subscribes to a [Future] and return its current state in an [AsyncSnapshot].
+///
+/// See also:
+///   * [Future]
+///   * [useValueListenable], [useListenable], [useAnimation]
+AsyncSnapshot<T> useFuture<T>(Future<T> future, {T initialData}) {
+  return Hook.use(_FutureHook(future, initialData: initialData));
 }
 
 class _FutureHook<T> extends Hook<AsyncSnapshot<T>> {
@@ -316,6 +553,15 @@ class _FutureStateHook<T> extends HookState<AsyncSnapshot<T>, _FutureHook<T>> {
   AsyncSnapshot<T> build(BuildContext context) {
     return _snapshot;
   }
+}
+
+/// Subscribes to a [Stream] and return its current state in an [AsyncSnapshot].
+///
+/// See also:
+///   * [Stream]
+///   * [useValueListenable], [useListenable], [useAnimation]
+AsyncSnapshot<T> useStream<T>(Stream<T> stream, {T initialData}) {
+  return Hook.use(_StreamHook(stream, initialData: initialData));
 }
 
 class _StreamHook<T> extends Hook<AsyncSnapshot<T>> {
@@ -410,6 +656,15 @@ class _StreamHookState<T> extends HookState<AsyncSnapshot<T>, _StreamHook<T>> {
       current.inState(ConnectionState.none);
 }
 
+/// A hook for side-effects
+///
+/// [useEffect] is called synchronously on every [HookWidget.build], unless
+/// [keys] is specified. In which case [useEffect] is called again only if
+/// any value inside [keys] as changed.
+void useEffect(VoidCallback Function() effect, [List keys]) {
+  Hook.use(_EffectHook(effect, keys));
+}
+
 class _EffectHook extends Hook<void> {
   final VoidCallback Function() effect;
 
@@ -468,6 +723,24 @@ class _EffectHookState extends HookState<void, _EffectHook> {
     print(hook.keys);
     properties.add(ObjectFlagProperty.has('dispose', disposer));
   }
+}
+
+/// Creates a [StreamController] automatically disposed.
+///
+/// See also:
+///   * [StreamController]
+///   * [useStream]
+StreamController<T> useStreamController<T>(
+    {bool sync = false,
+    VoidCallback onListen,
+    VoidCallback onCancel,
+    List keys}) {
+  return Hook.use(_StreamControllerHook(
+    onCancel: onCancel,
+    onListen: onListen,
+    sync: sync,
+    keys: keys,
+  ));
 }
 
 class _StreamControllerHook<T> extends Hook<StreamController<T>> {
@@ -532,6 +805,52 @@ class _StreamControllerHookState<T>
   @override
   void dispose() {
     _controller.close();
+    super.dispose();
+  }
+}
+
+/// Creates a [ValueNotifier] automatically disposed.
+///
+/// As opposed to `useState`, this hook do not subscribes to [ValueNotifier].
+/// This allows a more granular rebuild.
+///
+/// See also:
+///   * [ValueNotifier]
+///   * [useValueListenable]
+ValueNotifier<T> useValueNotifier<T>([T intialData, List keys]) {
+  return Hook.use(_ValueNotifierHook(
+    initialData: intialData,
+    keys: keys,
+  ));
+}
+
+class _ValueNotifierHook<T> extends Hook<ValueNotifier<T>> {
+  final T initialData;
+
+  const _ValueNotifierHook({List keys, this.initialData}) : super(keys: keys);
+
+  @override
+  _UseValueNotiferHookState<T> createState() => _UseValueNotiferHookState<T>();
+}
+
+class _UseValueNotiferHookState<T>
+    extends HookState<ValueNotifier<T>, _ValueNotifierHook<T>> {
+  ValueNotifier<T> notifier;
+
+  @override
+  void initHook() {
+    super.initHook();
+    notifier = ValueNotifier(hook.initialData);
+  }
+
+  @override
+  ValueNotifier<T> build(BuildContext context) {
+    return notifier;
+  }
+
+  @override
+  void dispose() {
+    notifier.dispose();
     super.dispose();
   }
 }