+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 0.2.1:
+
+- NEW: `useValueNotifier`, which creates a `ValueNotifier` similarly to `useState`. But without listening it.
+This can be useful to have a more granular rebuild when combined to `useValueListenable`.
+- NEW: `useContext`, which exposes the `BuildContext` of the currently building `HookWidget`. 
+
 ## 0.2.0:
 
 - Made all existing hooks as static functions, and removed `HookContext`. The migration is as followed:
+
 ```dart
 Widget build(HookContext context) {
     final state = context.useState(0);
@@ -8,6 +15,7 @@ Widget build(HookContext context) {
 ```
 
 becomes:
+
 ```dart
 Widget build(BuildContext context) {
     final state = useState(0);
@@ -15,6 +23,7 @@ Widget build(BuildContext context) {
 ```
 
 - Introduced keys for hooks and applied them to hooks where it makes sense.
+- Added `useReducer` for complex state. It is similar to `useState` but is being managed by a `Reducer` and can only be changed by dispatching an action.
 - fixes a bug where hot-reload without using hooks throwed an exception
 
 ## 0.1.0: